DIY S3 client
The old version used Rusoto, which relied on the Tokio
executor. Because this project is based on async-std, I had to write a
wrapper around all of the Tokio futures. It was a source of a few
bugs, so I wrote a new client based on Surf. I'm still pulling into
Rusoto as a dependency to do the authorization headers, but Tokio is
never involved.