Upgrade sqlx to 0.5.x
This was a bit of an adventure. Version 0.3 threaded lifetimes through
much of the code, which made it quite the headache to upgrade until I
found a GitHub comment that mentioned `DerefMut` impls to
`PgConnection`. That, plus reborrowing, allowed me to pass the
connections along in a fashion similar to before.