Clean up error handling
`tide::Error` is now a wrapper around `anyhow::Error`, and implements
`From`, so this removes the information-discarding wrapper I'd created
before. This also adds a middleware that extracts any errors that
reach the top level, logs them, and then reports them to Sentry. This
should make debugging much easier in the future.