Report logged and swallowed errors to Sentry (#28)
There are a number of exceptions that are logged with `log::error!`,
but those only show the top level error, making it impossible to see
the underlying cause. Reporting the errors with
`sentry::capture_error` (and `capture_anyhow` as appropriate) should
reveal the underlying causes.

This change tries to avoid capturing any errors that are propagated
upwards, to avoid sending duplicate errors into Sentry.